Fix crash in column editor and rename component placeholder to error state
