+++ packages/client/src/components/app/ComponentErrorState.svelte
